+++ scripts/fiddle-fork.sh
@@ -78,8 +78,12 @@ forkedOnDate=$(date +"%m-%d-%y";)
     cp -rf "../fiddles/${type}/${sourceFiddle}" "../fiddles/${type}/${targetFiddle}" || exit 89
 
     case ${type} in
-        'extjs'|'php'|'jquery'|'three'|'dojo'|'node'|'tween'|'chrome'|'svg')
+        'extjs'|'php'|'jquery'|'three'|'dojo'|'node'|'tween'|'chrome')
             $(cd bin; ./house-substr.sh ${sourceFiddle} ${targetFiddle} "../../fiddles/${type}/${targetFiddle}/app.js";) || exit 90
+            $(cd bin; ./house-substr.sh ${sourceFiddle} ${targetFiddle} "../../fiddles/${type}/${targetFiddle}/index.html";) || exit 91
+            $(cd bin; ./house-substr.sh ${sourceFiddle} ${targetFiddle} "../../fiddles/${type}/${targetFiddle}/README.markdown";) || exit 92
+            ;;
+        'svg')
             $(cd bin; ./house-substr.sh ${sourceFiddle} ${targetFiddle} "../../fiddles/${type}/${targetFiddle}/index.html";) || exit 91
             $(cd bin; ./house-substr.sh ${sourceFiddle} ${targetFiddle} "../../fiddles/${type}/${targetFiddle}/README.markdown";) || exit 92
             ;;
@@ -108,7 +112,7 @@ forkedOnDate=$(date +"%m-%d-%y";)
     fi
 
     case ${type} in
-        'extjs'|'php'|'jquery'|'three'|'dojo'|'node'|'tween')
+        'extjs'|'php'|'jquery'|'three'|'dojo'|'node'|'tween'|'svg')
             ./fiddle-index.sh ${type} || exit 95;
             ;;
         'chrome')